+++ src/Convert/ScalarConverter.php
@@ -6,17 +6,22 @@ class ScalarConverter implements TypeConverter
 {
     public function fromGraphQL($value): string
     {
-        if (! is_string($value)) {
-            throw new \InvalidArgumentException('Expected string, got ' . gettype($value));
-        }
-
-        return $value;
+        return $this->toString($value);
     }
 
     public function toGraphQL($value): string
     {
+        return $this->toString($value);
+    }
+
+    /**
+     * @param mixed $value Should be string
+     */
+    protected function toString($value): string
+    {
         if (! is_string($value)) {
-            throw new \InvalidArgumentException('Expected string, got ' . gettype($value));
+            $notString = gettype($value);
+            throw new \InvalidArgumentException("Expected string, got {$notString}");
         }
 
         return $value;