+++ src/Convert/StringConverter.php
@@ -6,14 +6,18 @@ class StringConverter implements TypeConverter
 {
     public function fromGraphQL($value): string
     {
-        if (! is_string($value)) {
-            throw new \InvalidArgumentException('Expected string, got ' . gettype($value));
-        }
-
-        return $value;
+        return $this->toString($value);
     }
 
     public function toGraphQL($value): string
+    {
+        return $this->toString($value);
+    }
+
+    /**
+     * @param mixed $value Should be string
+     */
+    protected function toString($value): string
     {
         if (! is_string($value)) {
             throw new \InvalidArgumentException('Expected string, got ' . gettype($value));