+++ src/Testing/MockRequest.php
@@ -2,15 +2,13 @@
 
 namespace Spawnia\Sailor\Testing;
 
-use stdClass;
-
 class MockRequest
 {
     public string $query;
 
-    public ?stdClass $variables;
+    public ?\stdClass $variables;
 
-    public function __construct(string $query, ?stdClass $variables = null)
+    public function __construct(string $query, ?\stdClass $variables = null)
     {
         $this->query = $query;
         $this->variables = $variables;